refactor: update eslint config, fix linter error for imports
